Improve efficiency of combinations function
I've added the ability for Pascal's triangle to dynamically expand as
higher values of nCr are requested.  This allows O(1) computation time
for combinations of n and r that are smaller than previously-computed
values, and significantly improved times for n and r that are larger but
still close to previously-computed n and r.